Move PublicDataFileFilter to open_humans/filters.py
