+++ open_humans/filters.py
@@ -1,8 +1,13 @@
 from django.utils.safestring import mark_safe
 
+from django_filters import CharFilter, FilterSet, MultipleChoiceFilter
 from django_filters.fields import DateRangeField
 from django_filters.filters import RangeFilter
-from django_filters.widgets import RangeWidget
+from django_filters.widgets import CSVWidget, RangeWidget
+
+from data_import.models import DataFile
+from private_sharing.utilities import (
+     get_source_labels_and_names_including_dynamic)
 
 
 class StartEndRangeWidget(RangeWidget):
@@ -61,3 +66,22 @@ class StartEndDateFromToRangeFilter(RangeFilter):
     """
 
     field_class = StartEndDateRangeField
+
+
+class PublicDataFileFilter(FilterSet):
+    """
+    A FilterSet that maps member_id and username to less verbose names.
+    """
+
+    created = StartEndDateFromToRangeFilter()
+    member_id = CharFilter(field_name='user__member__member_id')
+    username = CharFilter(field_name='user__username')
+    source = MultipleChoiceFilter(
+        choices=get_source_labels_and_names_including_dynamic,
+        widget=CSVWidget())
+    # don't filter by source if no sources are specified; this improves speed
+    source.always_filter = False
+
+    class Meta:  # noqa: D101
+        model = DataFile
+        fields = ('created', 'source', 'username', 'member_id')
